+++ src/Presentation/SmartStore.Web/Administration/Models/Settings/SearchSettingsModel.cs
@@ -55,7 +55,7 @@ namespace SmartStore.Admin.Models.Settings
         public ProductSortingEnum DefaultSortOrder { get; set; }
         public SelectList AvailableSortOrderModes { get; set; }
 
-        public CommonFacetSettingsModel CategoryFacet { get; set; }
+		public CommonFacetSettingsModel CategoryFacet { get; set; }
 		public CommonFacetSettingsModel BrandFacet { get; set; }
 		public CommonFacetSettingsModel PriceFacet { get; set; }
 		public CommonFacetSettingsModel RatingFacet { get; set; }
@@ -78,6 +78,9 @@ namespace SmartStore.Admin.Models.Settings
 
 		[SmartResourceDisplayName("Common.DisplayOrder")]
 		public int DisplayOrder { get; set; }
+
+		[SmartResourceDisplayName("Admin.Configuration.Settings.Search.IncludeNotAvailable")]
+		public bool IncludeNotAvailable { get; set; }
 
 		public IList<CommonFacetSettingsLocalizedModel> Locales { get; set; }
 	}